requirements.txt
fastapi
uvicorn
python-multipart
pillow
torch
torchvision
transformers
requests
python-dotenv
pydantic
numpy
aiofiles

# torch
# accelerate
# opencv-python
# spaces
# pillow
# numpy
# timm
# kornia
# prettytable
# typing
# scikit-image
# huggingface_hub
# transformers>=4.39.1
# gradio
# gradio_imageslider
# loadimg>=0.1.1
# einops